Always run processResources
Since the build timestamp will always change, processResources always
has to run to filter messages.properties with the correct build time.
Luckily, Gradle has an API for configuring when a task output is up to
date. This is faster than forcing a clean, since we don't have to delete
all the resource files. It also avoids updating anything but the file
that changed.