Fix SQL issue with View Indiv. Pat. Sx/Dx
When selecting either Sypmtom or Diagnosis for a pie
chart, the click through was causing an SQL exception
due to the SQL alias in the WHERE clause.

Add a filterBeanId of the corresponding Id that is
used for filtering.  Confirm that all other existing
functionality also works: Details, Crosstab, Pie
Charts.